working on services for a server related dashboard

+++ lib/commands/server_halt.ts
@@ -2,6 +2,7 @@
 import error from "../utilities/error.js";
 import file from "../utilities/file.js";
 import log from "../utilities/log.js";
+import server from "../transmit/server.js";
 import vars from "../utilities/vars.js";
 
 // 1. turn off active servers
@@ -10,11 +11,15 @@ import vars from "../utilities/vars.js";
 // 4. delete vars.sockets[server]
 // 5. delete server from vars.server
 // 6. remove server's directory
-// 7. remove server from the config.json file
-// 8. call the callback
+// 7. modify the server
+// 8. remove server from the config.json file
+// 9. call the callback
 
-const delete_server = function commands_deleteServer(name:string, callback:() => void):void {
-    const terminate:boolean = (vars.command === "delete_server");
+const server_halt = function commands_serverHalt(server:server, action:type_halt_action, callback:() => void):void {
+    const terminate:boolean = (vars.command === "delete_server"),
+        name:string = (typeof server.modification_name === "string")
+            ? server.modification_name
+            : server.name;
     if (vars.servers[name] === undefined) {
         error([
             `Server with name ${name} does not exist.`,
@@ -25,12 +30,17 @@ const delete_server = function commands_deleteServer(name:string, callback:() =>
             path_name:string = `${vars.path.project}servers${vars.sep + name + vars.sep}`,
             flags:store_flag = {
                 config: false,
-                remove: false
+                remove: (action === "destroy")
+                    ? false
+                    : true,
+                restart: (action === "modify")
+                    ? false
+                    : true
             },
-            complete = function commands_deleteServer_complete(flag:"config"|"destroy"|"remove"):void {
+            complete = function commands_serverHalt_complete(flag:"config"|"remove"|"restart"):void {
                 flags[flag] = true;
-                if (flags.config === true && flags.remove === true) {
-                    if (vars.command === "delete_server") {
+                if (flags.config === true && flags.remove === true && flags.restart === true) {
+                    if (terminate === true) {
                         log([
                             `Server ${vars.text.cyan + name + vars.text.none} ${vars.text.angry}deleted${vars.text.none}.`,
                             `${vars.text.angry}*${vars.text.none} config.json file updated.`,
@@ -41,26 +51,32 @@ const delete_server = function commands_deleteServer(name:string, callback:() =>
                             ""
                         ], true);
                     } else {
+                        // 9. call the callback
                         callback();
                     }
                 }
             },
             file_remove:file_remove = {
-                callback: function commands_deleteServer_remove():void {
+                callback: function commands_serverHalt_remove():void {
                     complete("remove");
                 },
                 error_terminate: terminate,
                 exclusions: [],
                 location: path_name
             };
-        if (vars.command !== "delete_server") {
+        // halt active server
+        if (terminate === false) {
             const sockets:websocket_client[] = vars.sockets[name];
             let index:number = (sockets === undefined)
                 ? 0
                 : sockets.length;
             // 1. turn off active servers
-            vars.store_server.open[name].close();
-            vars.store_server.secure[name].close();
+            if (vars.store_server.open[name] !== undefined) {
+                vars.store_server.open[name].close();
+            }
+            if (vars.store_server.secure[name] !== undefined) {
+                vars.store_server.secure[name].close();
+            }
             // 2. kill all sockets on the server
             if (index > 0) {
                 do {
@@ -74,13 +90,23 @@ const delete_server = function commands_deleteServer(name:string, callback:() =>
             // 4. delete vars.sockets[server]
             delete vars.sockets[name];
         }
-        // 5. delete server from vars.server
-        delete vars.servers[name];
-        // 6. remove server's directory
-        file.remove(file_remove);
-        // 7. remove server from the config.json file
+        if (action === "destroy") {
+            // 5. delete server from vars.server
+            delete vars.servers[name];
+            // 6. remove server's directory
+            file.remove(file_remove);
+        }
+        // 7. restart the server
+        if (action === "modify") {
+            const modName:string = server.modification_name;
+            if (modName !== server.name && vars.servers[modName] !== undefined) {
+                delete server.modification_name;
+                vars.servers[server.name] = vars.servers[modName];
+            }
+        }
+        // 8. modify the config.json file
         file.write({
-            callback: function commands_deleteServer_readConfig_callback():void {
+            callback: function commands_serverHalt_readConfig_callback():void {
                 complete("config");
             },
             contents: JSON.stringify(vars.servers),
@@ -90,4 +116,4 @@ const delete_server = function commands_deleteServer(name:string, callback:() =>
     }
 };
 
-export default delete_server;
+export default server_halt;
